+++ apps/meteor/app/lib/server/functions/processWebhookMessage.ts
@@ -24,6 +24,7 @@ type Payload = {
 	bot?: IMessage['bot'];
 	groupable?: IMessage['groupable'];
 	tmid?: IMessage['tmid'];
+	customFields?: IMessage['customFields'];
 };
 
 type DefaultValues = {
@@ -110,6 +111,7 @@ export const processWebhookMessage = async function (
 			bot: messageObj.bot,
 			groupable: messageObj.groupable !== undefined ? messageObj.groupable : false,
 			tmid: messageObj.tmid,
+			customFields: messageObj.customFields,
 		};
 
 		if ((messageObj.icon_url && messageObj.icon_url.length > 0) || (messageObj.avatar && messageObj.avatar.length > 0)) {