+++ apps/meteor/tests/e2e/page-objects/fragments/account-sidenav.ts
@@ -10,4 +10,8 @@ export class AccountSidenav {
 	get linkTokens(): Locator {
 		return this.page.locator('.flex-nav [href="/account/tokens"]');
 	}
+
+	get linkSecurity(): Locator {
+		return this.page.locator('.flex-nav [href="/account/security"]');
+	}
 }