+++ apps/meteor/tests/e2e/page-objects/omnichannel-settings.ts
@@ -7,6 +7,22 @@ export class OmnichannelSettings {
 		this.page = page;
 	}
 
+	get labelLivechatLogo(): Locator {
+		return this.page.locator('//label[@title="Assets_livechat_widget_logo"]');
+	}
+
+	get imgLivechatLogoPreview(): Locator {
+		return this.page.locator('//label[@title="Assets_livechat_widget_logo"]/following-sibling::span >> role=img[name="Asset preview"]');
+	}
+
+	get inputLivechatLogo(): Locator {
+		return this.page.locator('//label[@title="Assets_livechat_widget_logo"]/following-sibling::span >> input[type="file"]');
+	}
+
+	get btnDeleteLivechatLogo(): Locator {
+		return this.page.locator('//label[@title="Assets_livechat_widget_logo"]/following-sibling::span >> role=button[name="Delete"]');
+	}
+
 	group(sectionName: string): Locator {
 		return this.page.locator(`[data-qa-section="${sectionName}"] h2 >> text="${sectionName}"`);
 	}