+++ packages/core-services/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rocket.chat/core-services
 
+## 0.3.12
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.7.1
+  - @rocket.chat/rest-typings@6.7.1
+  - @rocket.chat/models@0.0.36
+  </details>
+
 ## 0.3.11
 
 ### Patch Changes